// src/enums/product-category.enum.ts
export enum ProductCategory {
    ELECTRONICS = 'ELECTRONICS',
    PHONES_AND_ACCESSORIES = 'PHONES_AND_ACCESSORIES',
    COMPUTERS_AND_IT = 'COMPUTERS_AND_IT',
    HOME_APPLIANCES = 'HOME_APPLIANCES',
    FASHION_MEN = 'FASHION_MEN',
    FASHION_WOMEN = 'FASHION_WOMEN',
    FASHION_KIDS = 'FASHION_KIDS',
    SHOES = 'SHOES',
    ACCESSORIES = 'ACCESSORIES',
    BEAUTY_AND_PERSONAL_CARE = 'BEAUTY_AND_PERSONAL_CARE',
    HEALTH_AND_WELLNESS = 'HEALTH_AND_WELLNESS',
    HOME_AND_FURNITURE = 'HOME_AND_FURNITURE',
    HOME_DECOR = 'HOME_DECOR',
    KITCHEN_AND_UTENSILS = 'KITCHEN_AND_UTENSILS',
    BABY_PRODUCTS = 'BABY_PRODUCTS',
    TOYS_AND_GAMES = 'TOYS_AND_GAMES',
    SPORTS_AND_FITNESS = 'SPORTS_AND_FITNESS',
    AUTOMOTIVE = 'AUTOMOTIVE',
    MOTORCYCLES_AND_PARTS = 'MOTORCYCLES_AND_PARTS',
    CONSTRUCTION_AND_TOOLS = 'CONSTRUCTION_AND_TOOLS',
    OFFICE_AND_STATIONERY = 'OFFICE_AND_STATIONERY',
}